ДЗ_9

+++ app/src/main/java/com/example/habittracker/domain/models/HabitDto.kt
@@ -6,7 +6,7 @@ data class HabitDto (
     val count: Int,
     val date: Int,
     val description: String,
-    val done_dates: List<Int>,
+    val done_dates: List<Long>,
     val frequency: Int,
     val priority: Int,
     val title: String,
